pyproject.toml
[project]
name = "arize-mcp"
version = "0.1.0"
description = "MCP server to access Arize AX data, query traces, and run experiments"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "arize>=8.0.0b2",
    "pandas>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
arize-mcp = "arize_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arize_mcp"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
